\name{MBDDiff-package}
\alias{MBDDiff-package}
\docType{package}
\title{
\packageTitle{MBDDiff}
}
\description{
\packageDescription{MBDDiff}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MBDDiff}
\packageIndices{MBDDiff}
}
\author{
\packageAuthor{MBDDiff}

Maintainer: \packageMaintainer{MBDDiff}
}
\references{
	https://github.com/Liuy12/MBDDiff
}
