\name{MBDDiff-package}
\alias{MBDDiff-package}
\alias{MBDDiff}
\docType{package}
\title{
\packageTitle{MBDDiff}
}
\description{
\packageDescription{MBDDiff}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MBDDiff}
\packageIndices{MBDDiff}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{MBDDiff}

Maintainer: \packageMaintainer{MBDDiff}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
